+++ src/main/java/com/java/jacksondemo/repository/entity/UserRepository.java
@@ -1,0 +1,8 @@
+package com.java.jacksondemo.repository.entity;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface UserRepository extends JpaRepository<UserEntity, String> {
+
+    UserEntity findByUserId(long userId);
+}