+++ Changelog.md
@@ -1,4 +1,5 @@
 ###Added
 
-- implement json naming strategy
+- implement Jackson custom serialization.
+- implement json naming strategy.
 - implement jackson json deserialization using builder pattern.